resolve package import
